sandbox: Split EmptyLedgerIT into two.
