+++ ledger/sandbox/src/test/suite/scala/com/digitalasset/platform/sandbox/services/completion/CompletionServiceWithEmptyLedgerIT.scala
@@ -7,15 +7,8 @@ import java.io.File
 import java.nio.file.Files
 import java.time.Duration
 
-import com.daml.api.util.DurationConversion
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.testing.utils.{MockMessages, SuiteResourceManagementAroundEach}
-import com.daml.ledger.api.v1.admin.config_management_service.{
-  ConfigManagementServiceGrpc,
-  GetTimeModelRequest,
-  SetTimeModelRequest,
-  TimeModel => ProtobufTimeModel
-}
 import com.daml.ledger.api.v1.admin.package_management_service.{
   PackageManagementServiceGrpc,
   UploadDarFileRequest,
@@ -34,16 +27,15 @@ import com.daml.ledger.api.v1.command_completion_service.{
 }
 import com.daml.ledger.api.v1.command_submission_service.CommandSubmissionServiceGrpc
 import com.daml.ledger.api.v1.ledger_offset.LedgerOffset
-import com.daml.ledger.participant.state.v1.TimeModel
 import com.daml.platform.ApiOffset
 import com.daml.platform.sandbox.SandboxBackend
 import com.daml.platform.sandbox.config.SandboxConfig
 import com.daml.platform.sandbox.services.TestCommands
-import com.daml.platform.sandbox.services.completion.EmptyLedgerIT._
+import com.daml.platform.sandbox.services.TimeModelHelpers.publishATimeModel
+import com.daml.platform.sandbox.services.completion.CompletionServiceWithEmptyLedgerIT._
 import com.daml.platform.sandboxnext.SandboxNextFixture
 import com.daml.platform.testing.StreamConsumer
 import com.google.protobuf.ByteString
-import com.google.protobuf.timestamp.Timestamp
 import io.grpc.Channel
 import org.scalatest.{AsyncWordSpec, Inspectors, Matchers}
 import scalaz.syntax.tag._
@@ -51,7 +43,7 @@ import scalaz.syntax.tag._
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import scala.concurrent.{ExecutionContext, Future}
 
-final class EmptyLedgerIT
+final class CompletionServiceWithEmptyLedgerIT
     extends AsyncWordSpec
     with Matchers
     with Inspectors
@@ -60,15 +52,15 @@ final class EmptyLedgerIT
     with TestCommands
     with SuiteResourceManagementAroundEach {
 
-  // Start with empty daml packages and a large configuration delay, such that we can test the API's behavior
-  // on an empty index
+  // Start with no DAML packages and a large configuration delay, such that we can test the API's
+  // behavior on an empty index.
   override protected def config: SandboxConfig =
     super.config.copy(
       damlPackages = List.empty,
       ledgerConfig = super.config.ledgerConfig.copy(
-        initialConfigurationSubmitDelay = Duration.ofDays(5)
+        initialConfigurationSubmitDelay = Duration.ofDays(5),
       ),
-      implicitPartyAllocation = false
+      implicitPartyAllocation = false,
     )
 
   "CommandCompletionService gives sensible ledger end on an empty ledger" in {
@@ -87,17 +79,6 @@ final class EmptyLedgerIT
     } yield {
       end.getOffset.value.absolute.get shouldBe ApiOffset.begin.toHexString
       completions shouldBe empty
-    }
-  }
-
-  "ConfigManagementService accepts a configuration when none is set" in {
-    val lid = ledgerId().unwrap
-    val completionService = CommandCompletionServiceGrpc.stub(channel)
-    for {
-      _ <- publishATimeModel(channel)
-      end <- completionService.completionEnd(CompletionEndRequest(lid))
-    } yield {
-      end.getOffset.value.absolute.get should not be ApiOffset.begin.toHexString
     }
   }
 
@@ -128,28 +109,7 @@ final class EmptyLedgerIT
   }
 }
 
-object EmptyLedgerIT {
-  private def publishATimeModel(channel: Channel)(implicit ec: ExecutionContext): Future[Unit] = {
-    val configService = ConfigManagementServiceGrpc.stub(channel)
-    for {
-      current <- configService.getTimeModel(GetTimeModelRequest())
-      generation = current.configurationGeneration
-      timeModel = TimeModel.reasonableDefault
-      _ <- configService.setTimeModel(
-        SetTimeModelRequest(
-          "config-submission",
-          Some(Timestamp(30, 0)),
-          generation,
-          Some(ProtobufTimeModel(
-            avgTransactionLatency =
-              Some(DurationConversion.toProto(timeModel.avgTransactionLatency)),
-            minSkew = Some(DurationConversion.toProto(timeModel.minSkew)),
-            maxSkew = Some(DurationConversion.toProto(timeModel.maxSkew))
-          ))
-        ))
-    } yield ()
-  }
-
+object CompletionServiceWithEmptyLedgerIT {
   private def uploadDarFile(darFile: File, channel: Channel): Future[UploadDarFileResponse] = {
     val packageService = PackageManagementServiceGrpc.stub(channel)
     val darContents = {
